Update index.md
Just adding a bit for the sake of seeing the various features of the site. Although, I would reconsider the tab just underneath your line "Yes, that was a futurama..." I'm just not sure what purpose it serves. Maybe it will become clearer. But the black on black does not help.